chore: move tracer to core
